norm changes fix

+++ Libft_Future_Repo/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: apereira <apereira@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 11:42:55 by apereira          #+#    #+#             */
-/*   Updated: 2022/10/24 11:44:52 by apereira         ###   ########.fr       */
+/*   Updated: 2022/10/24 14:45:38 by apereira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
